+++ sdk/include/opentelemetry/sdk/configuration/configuration_parser.h
@@ -36,6 +36,7 @@
 #include "opentelemetry/sdk/configuration/double_array_attribute_value_configuration.h"
 #include "opentelemetry/sdk/configuration/double_attribute_value_configuration.h"
 #include "opentelemetry/sdk/configuration/drop_aggregation_configuration.h"
+#include "opentelemetry/sdk/configuration/event_to_span_event_bridge_log_record_processor_configuration.h"
 #include "opentelemetry/sdk/configuration/exemplar_filter.h"
 #include "opentelemetry/sdk/configuration/explicit_bucket_histogram_aggregation_configuration.h"
 #include "opentelemetry/sdk/configuration/extension_log_record_exporter_configuration.h"
@@ -165,6 +166,10 @@ public:
       const std::unique_ptr<DocumentNode> &node) const;
 
   std::unique_ptr<SimpleLogRecordProcessorConfiguration> ParseSimpleLogRecordProcessorConfiguration(
+      const std::unique_ptr<DocumentNode> &node) const;
+
+  std::unique_ptr<EventToSpanEventBridgeLogRecordProcessorConfiguration>
+  ParseEventToSpanEventBridgeLogRecordProcessorConfiguration(
       const std::unique_ptr<DocumentNode> &node) const;
 
   std::unique_ptr<ExtensionLogRecordProcessorConfiguration>